requirements.txt
streamlit
langgraph>=0.2.73
langgraph-checkpoint
langgraph-checkpoint-sqlite
langchain
langchain-core>=0.2.43
langchain-community
langchain-groq
langchain-text-splitters
langchain-mcp-adapters
pypdf
python-docx
python-dotenv
aiosqlite
SQLAlchemy
psycopg2-binary
pgvector
sentence-transformers
mcp
ddgs
requests
pandas
numpy
pillow
# Rebuild forced: 2026-03-10 23:45
fastmcp
